refactor client endpoint
